this commit will move the database connection to the server, correct some minor reference errors, and fix react-router syntax; front/back/database connections render correctly
