+++ client/index.js
@@ -1,11 +1,12 @@
 import React from 'react';
-import ReactDOM from 'react-dom/client';
+import { render } from 'react-dom'
+import { BrowserRouter } from 'react-router-dom';
 import './stylesheets/styles.scss';
-import App from './src/App.jsx';
+import App from './App';
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  <React.StrictMode>
+render(
+  <BrowserRouter>
     <App />
-  </React.StrictMode>
+  </BrowserRouter>,
+  document.getElementById('root'),
 );
