+++ package.json
@@ -22,7 +22,8 @@
     "react-dom": "^17.0.2",
     "react-hot-loader": "^4.13.1",
     "react-router": "^6.4.4",
-    "react-router-dom": "^6.4.4"
+    "react-router-dom": "^6.4.4",
+    "sass": "^1.56.2"
   },
   "nodemonConfig": {
     "ignore": [
